pyproject.toml
[tool.poetry]
name = "hw1-mlops"
version = "0.1.0"
description = "MLOps Homework Service"
authors = ["Student <vivasilevskiy_1@edu.hse.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
grpcio = "^1.59.0"
grpcio-tools = "^1.59.0"
protobuf = "^4.25.0"
scikit-learn = "^1.3.0"
pandas = "^2.1.0"
clearml = "^1.13.0"
dvc = {extras = ["s3"], version = "^3.30.0"}
python-multipart = "^0.0.6"
streamlit = "^1.28.0"
pydantic-settings = "^2.1.0"
boto3 = "^1.29.0"
s3fs = "^2023.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"